Add Bulma styling library
I had three options that I knew of:

- Install Bulma directly
  - Done this, [following instructions][1]
- Install vue-bulma
- Install Buefy

The further down that list you go, the more technical dependencies you hit.

I looked at Buefy, which is a full Vue-type implementation of Bulma. However,
for example, there are two-month old pull requests on that project. Buefy
would need to stay alive for the duration of this app's life.

Using vue-bulma is a good utility set of repos, but again there's the extra
layer of technical debt which could affect this project long-term.

I'm betting fully on VueJS, of course, and on Bulma.

[1]: https://www.bleedingedgepress.com/book_excerpts/01E9D1/creating_interfaces_with_bulma_sample.pdf